+++ certbot-dns-google/certbot_dns_google/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_google.dns_google` plugin.
